refactor(models): inline 8 ML enums into ml_tables.py; delete tables.py shim
backend/models/tables.py was a pure re-export shim of unified_models plus 8
ML-only enums (ModelTypeEnum, ModelStageEnum, FeatureTypeEnum, ComputeModeEnum,
FeatureStatusEnum, DriftTypeEnum, AlertSeverityEnum, ModelHealthEnum) that had
exactly one consumer: backend/ml/ml_tables.py.

- Inline those 8 enums directly into ml_tables.py.
- Rewrite ml_tables.py to pull Base from unified_models (canonical).
- Delete backend/models/tables.py.

models/__init__.py never re-exported tables.py, so no init change needed.
Smoke import (`backend.ml.ml_tables.ModelTypeEnum.SKLEARN.value`) passes.

Part of Tier B1 (docs/ANALYSIS_2026-05-14.md).
